requirements.txt
pytesseract
Pillow
gradio
regex
